Implement Ability to have difference timeline icons for experience(s)
Implement Ability to have difference timeline icons for experience. The default icon (currently angular) can be changed in Experience.js in the VerticalTimelineElement component by modifying the ternary operator for the icon to set a default icon if no icon is supplied from res_primaryLanguage.json and res_secondaryLanguage.json

Add a new key called and value to the relevant experience entry, specifying the Font Awesome 5 icon to use for this experience. e.g
 "timelineIcon": "fas fa-landmark",